backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
snscrape==0.7.0.20230622
textblob==0.17.1
vaderSentiment==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
